Add linking tests; consolidate test code
