+++ test/validating/validating.test.ts
@@ -1,27 +1,7 @@
 import { expect, test } from "vitest";
-import { EmptyFileSystem, type LangiumDocument } from "langium";
 import { expandToString as s } from "langium/generate";
-import { parseHelper } from "langium/test";
 import type { Diagnostic } from "vscode-languageserver-types";
-import { createTypicalServices } from "../../src/language/typical-module.js";
-import { Schema, isSchema } from "../../src/language/generated/ast.js";
-
-const init = async (files?: Record<string, string>) => {
-  const services = createTypicalServices(EmptyFileSystem);
-  const doParse = parseHelper<Schema>(services.Typical);
-  if (files) {
-    for (const [uri, content] of Object.entries(files)) {
-      services.shared.workspace.IndexManager.updateContent(
-        await doParse(content, {
-          validation: true,
-          documentUri: uri,
-        })
-      );
-    }
-  }
-  const parse = (input: string) => doParse(input, { validation: true });
-  return { services, parse };
-};
+import { init, expectDocumentValid } from "../utils";
 
 test("check no errors", async () => {
   const { parse } = await init();
@@ -54,7 +34,7 @@ test("missing local type is invalid", async () => {
   );
 });
 
-test.only("missing imported type is invalid", async () => {
+test("missing imported type is invalid", async () => {
   const { parse } = await init({
     "coords.t": "struct Position { x: S64 = 0, y: S64 = 1 }",
   });
@@ -74,23 +54,6 @@ test.only("missing imported type is invalid", async () => {
       `)
   );
 });
-
-function expectDocumentValid(document: LangiumDocument) {
-  expect(
-    (document.parseResult.parserErrors.length &&
-      s`
-        Parser errors:
-          ${document.parseResult.parserErrors
-            .map((e) => e.message)
-            .join("\n  ")}
-    `) ||
-      (document.parseResult.value === undefined &&
-        `ParseResult is 'undefined'.`) ||
-      (!isSchema(document.parseResult.value) &&
-        `Root AST object is a ${document.parseResult.value.$type}, expected a '${Schema}'.`) ||
-      undefined
-  ).toBeUndefined();
-}
 
 function diagnosticToString(d: Diagnostic) {
   return `[${d.range.start.line}:${d.range.start.character}..${d.range.end.line}:${d.range.end.character}]: ${d.message}`;
